decompose Planning component
