+++ client/src/components/_atoms/DogsBubbles/DogsBubbles.scss
