+++ client/src/components/_atoms/DogsBubbles/DogsBubbles.tsx
@@ -2,18 +2,16 @@ import type React from "react";
 import "./DogsBubbles.scss";
 import { useImageUrl } from "@/hooks/useImageUrl";
 import type { Dog } from "@/types/Dog";
+import { useNavigate } from "react-router-dom";
 
 interface DogBubblesProps {
 	dogs: Dog[];
 	maxSize: number;
-	onDogClick?: (dog: Dog) => void;
 }
 
-const DogBubbles: React.FC<DogBubblesProps> = ({
-	dogs,
-	maxSize,
-	onDogClick,
-}) => {
+const DogBubbles: React.FC<DogBubblesProps> = ({ dogs, maxSize }) => {
+	const navigate = useNavigate();
+
 	if (!dogs || dogs.length === 0) {
 		return (
 			<div className="participants-wrapper">
@@ -25,9 +23,7 @@ const DogBubbles: React.FC<DogBubblesProps> = ({
 	}
 
 	const handleDogClick = (dog: Dog) => {
-		if (onDogClick) {
-			onDogClick(dog);
-		}
+		navigate(`/dog/${dog.id}`);
 	};
 
 	return (
