+++ client/src/components/_atoms/OwnerBubble/OwnerBubble.tsx
