+++ client/src/components/_atoms/TrainerBubble/TrainerBubble.tsx
