+++ client/src/types/Event.d.ts
@@ -24,6 +24,21 @@ export interface Event {
 	services: ServiceType[];
 }
 
+export interface CalendarEvent {
+	id: string;
+	title: string;
+	start: Date;
+	end: Date;
+	description: string;
+	extendedProps: {
+		group_max_size: number;
+		location: Location;
+		price: number;
+		dogs: Dog[];
+		services: ServiceType[];
+	};
+}
+
 export interface Participation {
 	dog: Dog;
 }
